backend/pyproject.toml
[project]
name = "docubrain-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pypdf (>=6.4.0,<7.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "qdrant-client (>=1.16.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "mcp (>=1.22.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
